requirements.txt
# Core API
fastapi==0.125.0
uvicorn==0.35.0
pydantic==2.11.7
python-multipart==0.0.21

# HTTP / async
aiohttp==3.12.15
aiofiles==25.1.0
httpx==0.28.1
requests==2.32.5

# LLM / narrative
openai==2.7.1
tenacity==9.1.2
lightrag-hku==1.4.9.10
numpy==1.26.4

# Music
mutagen==1.47.0
pydub==0.25.1

# Parsing / scraping / analysis
beautifulsoup4==4.13.5
PyYAML==6.0.2
PyMuPDF==1.26.6
pandas==2.3.2
python-docx==1.2.0
python-pptx==1.0.2
tqdm==4.67.1
googlesearch-python==1.3.0
baidusearch==1.0.3

# Logging
colorama==0.4.6
